+++ app/src/main/java/com/example/gogreen/data/StartChargingRequest.kt
@@ -3,5 +3,5 @@ package com.example.gogreen.data
 data class StartChargingRequest(
     val station_id: String,
     val my_wallet_address: String,
-    val required_energy: Int
+    val required_energy: Double
 )
