+++ app/src/main/java/com/example/gogreen/data/StartChargingResponse.kt
@@ -3,11 +3,11 @@ package com.example.gogreen.data
 data class StartChargingResponse(
     val status: Boolean = true,
     val message: String? = "",
-    val energyConsumed: Int? = 0,
-    val price: Int? = 0,
-    val requiredEnergy: Int? = 0,
-    val timePassed: Int? = 0,
-    val totalTime: Int? = 0,
-    val transferRate: Int? = 0,
-    val totalPrice: Int? = 0
+    val energyConsumed: Double? = 0.0,
+    val price: Double? = 0.0,
+    val requiredEnergy: Double? = 0.0,
+    val timePassed: Double? = 0.0,
+    val totalTime: Double? = 0.0,
+    val transferRate: Double? = 0.0,
+    val totalPrice: Double? = 0.0
 )
